[feat/#74] dto 수정
