+++ app/src/main/java/com/poti/android/data/remote/datasource/PartyRemoteDataSource.kt
@@ -4,7 +4,7 @@ import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.request.party.CreatePartyRequestDto
 import com.poti.android.data.remote.dto.request.party.PartyJoinRequestDto
 import com.poti.android.data.remote.dto.response.artist.ArtistSearchListResponseDto
-import com.poti.android.data.remote.dto.response.history.GroupBuyPostSaleDto
+import com.poti.android.data.remote.dto.response.history.RecruiterDetailDto
 import com.poti.android.data.remote.dto.response.party.CreatePartyResponseDto
 import com.poti.android.data.remote.dto.response.party.MyRecruitListDto
 import com.poti.android.data.remote.dto.response.party.PartyDetailResponseDto
@@ -45,6 +45,6 @@ class PartyRemoteDataSource @Inject constructor(
     suspend fun getMyRecruitList(status: String): BaseResponse<MyRecruitListDto> =
         partyService.getMyRecruitList(status)
 
-    suspend fun getPostSale(postId: Long): BaseResponse<GroupBuyPostSaleDto> =
-        partyService.getPostSale(postId)
+    suspend fun getRecruitDetail(postId: Long): BaseResponse<RecruiterDetailDto> =
+        partyService.getRecruitDetail(postId)
 }
