+++ app/src/main/java/com/poti/android/data/remote/service/PartyService.kt
@@ -4,7 +4,7 @@ import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.request.party.CreatePartyRequestDto
 import com.poti.android.data.remote.dto.request.party.PartyJoinRequestDto
 import com.poti.android.data.remote.dto.response.artist.ArtistSearchListResponseDto
-import com.poti.android.data.remote.dto.response.history.GroupBuyPostSaleDto
+import com.poti.android.data.remote.dto.response.history.RecruiterDetailDto
 import com.poti.android.data.remote.dto.response.party.CreatePartyResponseDto
 import com.poti.android.data.remote.dto.response.party.MyRecruitListDto
 import com.poti.android.data.remote.dto.response.party.PartyDetailResponseDto
@@ -59,7 +59,7 @@ interface PartyService {
     ): BaseResponse<MyRecruitListDto>
 
     @GET("/api/v1/posts/sale/{postId}")
-    suspend fun getPostSale(
+    suspend fun getRecruitDetail(
         @Path("postId") postId: Long,
-    ): BaseResponse<GroupBuyPostSaleDto>
+    ): BaseResponse<RecruiterDetailDto>
 }
