+++ app/src/main/java/com/poti/android/domain/repository/PartyRepository.kt
@@ -41,5 +41,5 @@ interface PartyRepository {
 
     suspend fun getMyRecruitList(status: String): Result<MyPartyList>
 
-    suspend fun getPostSale(postId: Long): Result<RecruiterDetail>
+    suspend fun getRecruitDetail(postId: Long): Result<RecruiterDetail>
 }
